+++ packages/code-studio/src/settings/SettingsMenu.tsx
@@ -29,10 +29,10 @@ import { type ServerConfigValues, type User, store } from '@deephaven/redux';
 import {
   BROADCAST_CHANNEL_NAME,
   BROADCAST_LOGOUT_MESSAGE,
-  makeMessage,
 } from '@deephaven/jsapi-utils';
 import { type PluginModuleMap } from '@deephaven/plugin';
 import { exportLogs, logHistory } from '@deephaven/log';
+import { makeMessage } from '@deephaven/utils';
 import FormattingSectionContent from './FormattingSectionContent';
 import LegalNotice from './LegalNotice';
 import SettingsMenuSection from './SettingsMenuSection';