+++ packages/iris-grid/src/IrisGrid.tsx
@@ -1831,6 +1831,12 @@ class IrisGrid extends Component<IrisGridProps, IrisGridState> {
     });
   }
 
+  clearAllAggregations(): void {
+    log.debug('Clearing all aggregations');
+
+    this.setState({ aggregationSettings: DEFAULT_AGGREGATION_SETTINGS });
+  }
+
   clearCrossColumSearch(): void {
     log.debug('Clearing cross-column search');
 
@@ -3212,7 +3218,7 @@ class IrisGrid extends Component<IrisGridProps, IrisGridState> {
 
   handleTableChanged(): void {
     const { model } = this.props;
-    // movedColumns update triggers metricCalculator update in the Grid component
+    // movedColumns reset triggers metricCalculator update in the Grid component
     this.setState({ movedColumns: model.initialMovedColumns });
   }
 
@@ -3648,6 +3654,7 @@ class IrisGrid extends Component<IrisGridProps, IrisGridState> {
 
     this.showAllColumns();
     this.clearAllFilters();
+    this.clearAllAggregations();
 
     this.startLoading(
       `Selecting distinct values in ${