+++ packages/iris-grid/src/IrisGridTheme.ts
@@ -152,7 +152,7 @@ export function createDefaultIrisGridTheme(): IrisGridThemeType {
     rowHeight: parseInt(IrisGridTheme['row-height'], 10) || 19, // IrisGrid test breaks without the fallback value
     columnWidth: 100,
     rowHeaderWidth: 0,
-    rowFooterWidth: 60,
+    rowFooterWidth: 90,
     columnHeaderHeight: parseInt(IrisGridTheme['header-height'], 10) || 30,
     filterBarHeight: 30, // includes 1px casing at bottom
     filterBarCollapsedHeight: 5, // includes 1px casing at bottom